+++ camera/CameraManager.swift
@@ -9,6 +9,11 @@
 import UIKit
 import AVFoundation
 import Photos
+import ImageIO
+import MobileCoreServices
+import Photos
+import CoreLocation
+
 public enum CameraState {
     case ready, accessDenied, noDeviceFound, notDetermined
 }
@@ -68,6 +73,25 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
         }
     }
     
+    /// Property to determine if manager should horizontally flip image took by front camera. Default value is false.
+    open var shouldFlipFrontCameraImage = false
+    
+    open var shouldKeepViewAtOrientationChanges = false
+    
+    /// Property to determine if manager should enable tap to focus on camera preview. Default value is true.
+    open var shouldEnableTapToFocus = true {
+        didSet {
+            focusGesture.isEnabled = shouldEnableTapToFocus
+        }
+    }
+    
+    /// Property to determine if manager should enable pinch to zoom on camera preview. Default value is true.
+    open var shouldEnablePinchToZoom = true {
+        didSet {
+            zoomGesture.isEnabled = shouldEnablePinchToZoom
+        }
+    }
+    
     /// The Bool property to determine if the camera is ready to use.
     open var cameraIsReady: Bool {
         get {
@@ -77,37 +101,31 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
     
     /// The Bool property to determine if current device has front camera.
     open var hasFrontCamera: Bool = {
-        let devices = AVCaptureDevice.devices(withMediaType: AVMediaTypeVideo)
-        for  device in devices!  {
-            let captureDevice = device as! AVCaptureDevice
-            if (captureDevice.position == .front) {
-                return true
-            }
-        }
-        return false
+        let frontDevices = AVCaptureDevice.videoDevices.filter { $0.position == .front }
+        return !frontDevices.isEmpty
     }()
     
     /// The Bool property to determine if current device has flash.
     open var hasFlash: Bool = {
-        let devices = AVCaptureDevice.devices(withMediaType: AVMediaTypeVideo)
-        for  device in devices!  {
-            let captureDevice = device as! AVCaptureDevice
-            if (captureDevice.position == .back) {
-                return captureDevice.hasFlash
-            }
-        }
-        return false
+        let hasFlashDevices = AVCaptureDevice.videoDevices.filter { $0.hasFlash }
+        return !hasFlashDevices.isEmpty
     }()
+    
+    /// Property to enable or disable switch animation
+    
+    open var animateCameraDeviceChange: Bool = true
     
     /// Property to change camera device between front and back.
     open var cameraDevice = CameraDevice.back {
         didSet {
             if cameraIsSetup {
                 if cameraDevice != oldValue {
+                    if animateCameraDeviceChange {
+                        _doFlipAnimation()
+                    }
                     _updateCameraDevice(cameraDevice)
                     _setupMaxZoomScale()
                     _zoom(0)
-                    _orientationChanged()
                 }
             }
         }
@@ -154,8 +172,14 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
     /// Property to check video recording file size when in progress
     open var recordedFileSize : Int64 { return movieOutput?.recordedFileSize ?? 0 }
     
+    //Properties to set focus and capture mode when tap to focus is used (_focusStart)
+    open var focusMode : AVCaptureFocusMode = .continuousAutoFocus
+    open var exposureMode: AVCaptureExposureMode = .continuousAutoExposure
+    
     
     // MARK: - Private properties
+    
+    fileprivate var locationManager = CameraLocationManager()
     
     fileprivate weak var embeddingView: UIView?
     fileprivate var videoCompletion: ((_ videoURL: URL?, _ error: NSError?) -> Void)?
@@ -163,22 +187,20 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
     fileprivate var sessionQueue: DispatchQueue = DispatchQueue(label: "CameraSessionQueue", attributes: [])
     
     fileprivate lazy var frontCameraDevice: AVCaptureDevice? = {
-        let devices = AVCaptureDevice.devices(withMediaType: AVMediaTypeVideo) as! [AVCaptureDevice]
-        return devices.filter{$0.position == .front}.first
+        return AVCaptureDevice.videoDevices.filter { $0.position == .front }.first
     }()
     
     fileprivate lazy var backCameraDevice: AVCaptureDevice? = {
-        let devices = AVCaptureDevice.devices(withMediaType: AVMediaTypeVideo) as! [AVCaptureDevice]
-        return devices.filter{$0.position == .back}.first
+        return AVCaptureDevice.videoDevices.filter { $0.position == .back }.first
     }()
     
     fileprivate lazy var mic: AVCaptureDevice? = {
         return AVCaptureDevice.defaultDevice(withMediaType: AVMediaTypeAudio)
     }()
     
-    open var previewLayer: AVCaptureVideoPreviewLayer?
     fileprivate var stillImageOutput: AVCaptureStillImageOutput?
     fileprivate var movieOutput: AVCaptureMovieFileOutput?
+    fileprivate var previewLayer: AVCaptureVideoPreviewLayer?
     fileprivate var library: PHPhotoLibrary?
     
     fileprivate var cameraIsSetup = false
@@ -189,13 +211,14 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
     fileprivate var maxZoomScale    = CGFloat(1.0)
     
     fileprivate var tempFilePath: URL = {
-        let tempPath = URL(fileURLWithPath: NSTemporaryDirectory()).appendingPathComponent("tempMovie").appendingPathExtension("mp4").absoluteString
+        let tempURL = URL(fileURLWithPath: NSTemporaryDirectory()).appendingPathComponent("tempMovie").appendingPathExtension("mp4")
+        let tempPath = tempURL.absoluteString
         if FileManager.default.fileExists(atPath: tempPath) {
             do {
                 try FileManager.default.removeItem(atPath: tempPath)
             } catch { }
         }
-        return URL(string: tempPath)!
+        return tempURL
     }()
     
     
@@ -249,16 +272,16 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
      :param: completion Completion block with the result of permission request
      */
     open func askUserForCameraPermission(_ completion: @escaping (Bool) -> Void) {
-        AVCaptureDevice.requestAccess(forMediaType: AVMediaTypeVideo, completionHandler: { (alowedAccess) -> Void in
+        AVCaptureDevice.requestAccess(forMediaType: AVMediaTypeVideo, completionHandler: { (allowedAccess) -> Void in
             if self.cameraOutputMode == .videoWithMic {
-                AVCaptureDevice.requestAccess(forMediaType: AVMediaTypeAudio, completionHandler: { (alowedAccess) -> Void in
-                    DispatchQueue.main.sync(execute: { () -> Void in
-                        completion(alowedAccess)
+                AVCaptureDevice.requestAccess(forMediaType: AVMediaTypeAudio, completionHandler: { (allowedAccess) -> Void in
+                    DispatchQueue.main.async(execute: { () -> Void in
+                        completion(allowedAccess)
                     })
                 })
             } else {
-                DispatchQueue.main.sync(execute: { () -> Void in
-                    completion(alowedAccess)
+                DispatchQueue.main.async(execute: { () -> Void in
+                    completion(allowedAccess)
                 })
                 
             }
@@ -302,6 +325,9 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
      */
     open func stopAndRemoveCaptureSession() {
         stopCaptureSession()
+        let oldAnimationValue = animateCameraDeviceChange
+        animateCameraDeviceChange = false
+        cameraDevice = .back
         cameraIsSetup = false
         previewLayer = nil
         captureSession = nil
@@ -310,6 +336,7 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
         mic = nil
         stillImageOutput = nil
         movieOutput = nil
+        animateCameraDeviceChange = oldAnimationValue
     }
     
     /**
@@ -325,22 +352,43 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
                 return
             }
             
-            if self.writeFilesToPhoneLibrary == true, let library = self.library  {
+            self._performShutterAnimation() {
+                guard let tempImage = UIImage(data: imageData) else {
+                    imageCompletion(nil, NSError())
+                    return
+                }
                 
-                
-                library.performChanges({
-                    PHAssetChangeRequest.creationRequestForAsset(from: UIImage(data: imageData)!)
-                }, completionHandler: { success, error in
-                    guard error != nil else {
+                let image: UIImage
+                if self.shouldFlipFrontCameraImage == true, self.cameraDevice == .front {
+                    guard let cgImage = tempImage.cgImage else {
+                        imageCompletion(nil, NSError())
                         return
                     }
-                    
-                    DispatchQueue.main.async(execute: {
-                        self._show(NSLocalizedString("Error", comment:""), message: (error?.localizedDescription)!)
+                    let flippedImage = UIImage(cgImage: cgImage, scale: tempImage.scale, orientation: .leftMirrored)
+                    image = flippedImage
+                } else {
+                    image = tempImage
+                }
+                
+                if self.writeFilesToPhoneLibrary == true, let library = self.library  {
+                    library.performChanges({
+                        let request = PHAssetChangeRequest.creationRequestForAsset(from: image)
+                        request.creationDate = Date()
+                        
+                        if let location = self.locationManager.latestLocation {
+                            request.location = location
+                        }
+                    }, completionHandler: { success, error in
+                        if let error = error {
+                            DispatchQueue.main.async(execute: {
+                                self._show(NSLocalizedString("Error", comment:""), message: error.localizedDescription)
+                            })
+                        }
                     })
-                })
+                }
+                
+                imageCompletion(image, nil)
             }
-            imageCompletion(UIImage(data: imageData), nil)
         }
     }
     
@@ -362,20 +410,18 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
         }
         
         sessionQueue.async(execute: {
-            self._getStillImageOutput().captureStillImageAsynchronously(from: self._getStillImageOutput().connection(withMediaType: AVMediaTypeVideo), completionHandler: { [unowned self] sample, error in
+            let stillImageOutput = self._getStillImageOutput()
+            stillImageOutput.captureStillImageAsynchronously(from: stillImageOutput.connection(withMediaType: AVMediaTypeVideo), completionHandler: { [weak self] sample, error in
                 
-                
-                guard error == nil else {
+                if let error = error {
                     DispatchQueue.main.async(execute: {
-                        self._show(NSLocalizedString("Error", comment:""), message: (error?.localizedDescription)!)
+                        self?._show(NSLocalizedString("Error", comment:""), message: error.localizedDescription)
                     })
                     imageCompletion(nil, error as NSError?)
                     return
                 }
                 
                 let imageData = AVCaptureStillImageOutput.jpegStillImageNSDataRepresentation(sample)
-                
-                
                 imageCompletion(imageData, nil)
                 
             })
@@ -421,7 +467,8 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
      :returns: Current flash mode: Off / On / Auto
      */
     open func changeFlashMode() -> CameraFlashMode {
-        flashMode = CameraFlashMode(rawValue: (flashMode.rawValue+1)%3)!
+        guard let newFlashMode = CameraFlashMode(rawValue: (flashMode.rawValue+1)%3) else { return flashMode }
+        flashMode = newFlashMode
         return flashMode
     }
     
@@ -431,13 +478,14 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
      :returns: Current quality mode: Low / Medium / High
      */
     open func changeQualityMode() -> CameraOutputQuality {
-        cameraOutputQuality = CameraOutputQuality(rawValue: (cameraOutputQuality.rawValue+1)%3)!
+        guard let newQuality = CameraOutputQuality(rawValue: (cameraOutputQuality.rawValue+1)%3) else { return cameraOutputQuality }
+        cameraOutputQuality = newQuality
         return cameraOutputQuality
     }
     
     // MARK: - AVCaptureFileOutputRecordingDelegate
     
-    open func capture(_ captureOutput: AVCaptureFileOutput!, didStartRecordingToOutputFileAt fileURL: URL!, fromConnections connections: [Any]!) {
+    open func capture(_ captureOutput: AVCaptureFileOutput?, didStartRecordingToOutputFileAt fileURL: URL?, fromConnections connections: [Any]?) {
         captureSession?.beginConfiguration()
         if flashMode != .off {
             _updateTorch(flashMode)
@@ -445,13 +493,14 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
         captureSession?.commitConfiguration()
     }
     
-    open func capture(_ captureOutput: AVCaptureFileOutput!, didFinishRecordingToOutputFileAt outputFileURL: URL!, fromConnections connections: [Any]!, error: Error!) {
+    open func capture(_ captureOutput: AVCaptureFileOutput?, didFinishRecordingToOutputFileAt outputFileURL: URL?, fromConnections connections: [Any]?, error: Error?) {
         _updateTorch(.off)
-        if (error != nil) {
+        if let error = error {
             _show(NSLocalizedString("Unable to save video to the iPhone", comment:""), message: error.localizedDescription)
-        } else {
+        }
+        else {
             
-            if writeFilesToPhoneLibrary {
+            if let outputFileURL = outputFileURL, writeFilesToPhoneLibrary {
                 
                 if PHPhotoLibrary.authorizationStatus() == .authorized {
                     saveVideoToLibrary(outputFileURL)
@@ -476,8 +525,8 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
                 
                 PHAssetChangeRequest.creationRequestForAssetFromVideo(atFileURL: fileURL)
             }, completionHandler: { success, error in
-                if (error != nil) {
-                    self._show(NSLocalizedString("Unable to save video to the iPhone.", comment:""), message: error!.localizedDescription)
+                if let error = error {
+                    self._show(NSLocalizedString("Unable to save video to the iPhone.", comment:""), message: error.localizedDescription)
                     self._executeVideoCompletionWithURL(nil, error: error as NSError?)
                 } else {
                     self._executeVideoCompletionWithURL(fileURL, error: error as NSError?)
@@ -487,11 +536,12 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
     }
     
     // MARK: - UIGestureRecognizerDelegate
+    fileprivate lazy var zoomGesture = UIPinchGestureRecognizer()
     
     fileprivate func attachZoom(_ view: UIView) {
-        let pinch = UIPinchGestureRecognizer(target: self, action: #selector(CameraManager._zoomStart(_:)))
-        view.addGestureRecognizer(pinch)
-        pinch.delegate = self
+        zoomGesture.addTarget(self, action: #selector(CameraManager._zoomStart(_:)))
+        view.addGestureRecognizer(zoomGesture)
+        zoomGesture.delegate = self
     }
     
     open func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
@@ -541,27 +591,157 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
         }
     }
     
+    // MARK: - UIGestureRecognizerDelegate
+    fileprivate lazy var focusGesture = UITapGestureRecognizer()
+    
+    fileprivate func attachFocus(_ view: UIView) {
+        focusGesture.addTarget(self, action: #selector(CameraManager._focusStart(_:)))
+        view.addGestureRecognizer(focusGesture)
+        focusGesture.delegate = self
+    }
+    
+    @objc fileprivate func _focusStart(_ recognizer: UITapGestureRecognizer) {
+        
+        let device: AVCaptureDevice?
+        
+        switch cameraDevice {
+        case .back:
+            device = backCameraDevice
+        case .front:
+            device = frontCameraDevice
+        }
+        
+        if let validDevice = device {
+            
+            //            if validDevice.isAdjustingFocus || validDevice.isAdjustingExposure || showingFocusRectangle {
+            //
+            //                return
+            //            }
+            
+            if let validPreviewLayer = previewLayer,
+                let view = recognizer.view
+            {
+                let pointInPreviewLayer = view.layer.convert(recognizer.location(in: view), to: validPreviewLayer)
+                let pointOfInterest = validPreviewLayer.captureDevicePointOfInterest(for: pointInPreviewLayer)
+                
+                do {
+                    try validDevice.lockForConfiguration()
+                    
+                    _showFocusRectangleAtPoint(pointInPreviewLayer, inLayer: validPreviewLayer)
+                    
+                    if validDevice.isFocusPointOfInterestSupported {
+                        validDevice.focusPointOfInterest = pointOfInterest;
+                    }
+                    
+                    if  validDevice.isExposurePointOfInterestSupported {
+                        validDevice.exposurePointOfInterest = pointOfInterest;
+                    }
+                    
+                    if validDevice.isFocusModeSupported(focusMode) {
+                        validDevice.focusMode = focusMode
+                    }
+                    
+                    if validDevice.isExposureModeSupported(exposureMode) {
+                        validDevice.exposureMode = exposureMode
+                    }
+                    
+                    validDevice.unlockForConfiguration()
+                }
+                catch let error {
+                    print(error)
+                }
+            }
+        }
+    }
+    
+    fileprivate var lastFocusRectangle:CAShapeLayer? = nil
+    
+    fileprivate func _showFocusRectangleAtPoint(_ focusPoint: CGPoint, inLayer layer: CALayer) {
+        
+        if let lastFocusRectangle = lastFocusRectangle {
+            
+            lastFocusRectangle.removeFromSuperlayer()
+            self.lastFocusRectangle = nil
+        }
+        
+        let size = CGSize(width: 75, height: 75)
+        let rect = CGRect(origin: CGPoint(x: focusPoint.x - size.width / 2.0, y: focusPoint.y - size.height / 2.0), size: size)
+        
+        let endPath = UIBezierPath(rect: rect)
+        endPath.move(to: CGPoint(x: rect.minX + size.width / 2.0, y: rect.minY))
+        endPath.addLine(to: CGPoint(x: rect.minX + size.width / 2.0, y: rect.minY + 5.0))
+        endPath.move(to: CGPoint(x: rect.maxX, y: rect.minY + size.height / 2.0))
+        endPath.addLine(to: CGPoint(x: rect.maxX - 5.0, y: rect.minY + size.height / 2.0))
+        endPath.move(to: CGPoint(x: rect.minX + size.width / 2.0, y: rect.maxY))
+        endPath.addLine(to: CGPoint(x: rect.minX + size.width / 2.0, y: rect.maxY - 5.0))
+        endPath.move(to: CGPoint(x: rect.minX, y: rect.minY + size.height / 2.0))
+        endPath.addLine(to: CGPoint(x: rect.minX + 5.0, y: rect.minY + size.height / 2.0))
+        
+        let startPath = UIBezierPath(cgPath: endPath.cgPath)
+        let scaleAroundCenterTransform = CGAffineTransform(translationX: -focusPoint.x, y: -focusPoint.y).concatenating(CGAffineTransform(scaleX: 2.0, y: 2.0).concatenating(CGAffineTransform(translationX: focusPoint.x, y: focusPoint.y)))
+        startPath.apply(scaleAroundCenterTransform)
+        
+        let shapeLayer = CAShapeLayer()
+        shapeLayer.path = endPath.cgPath
+        shapeLayer.fillColor = UIColor.clear.cgColor
+        shapeLayer.strokeColor = UIColor(red:1, green:0.83, blue:0, alpha:0.95).cgColor
+        shapeLayer.lineWidth = 1.0
+        
+        layer.addSublayer(shapeLayer)
+        lastFocusRectangle = shapeLayer
+        
+        CATransaction.begin()
+        
+        CATransaction.setAnimationDuration(0.2)
+        CATransaction.setAnimationTimingFunction(CAMediaTimingFunction(name: kCAMediaTimingFunctionEaseOut))
+        
+        CATransaction.setCompletionBlock() {
+            if shapeLayer.superlayer != nil {
+                shapeLayer.removeFromSuperlayer()
+                self.lastFocusRectangle = nil
+            }
+        }
+        
+        let appearPathAnimation = CABasicAnimation(keyPath: "path")
+        appearPathAnimation.fromValue = startPath.cgPath
+        appearPathAnimation.toValue = endPath.cgPath
+        shapeLayer.add(appearPathAnimation, forKey: "path")
+        
+        let appearOpacityAnimation = CABasicAnimation(keyPath: "opacity")
+        appearOpacityAnimation.fromValue = 0.0
+        appearOpacityAnimation.toValue = 1.0
+        shapeLayer.add(appearOpacityAnimation, forKey: "opacity")
+        
+        let disappearOpacityAnimation = CABasicAnimation(keyPath: "opacity")
+        disappearOpacityAnimation.fromValue = 1.0
+        disappearOpacityAnimation.toValue = 0.0
+        disappearOpacityAnimation.beginTime = CACurrentMediaTime() + 0.8
+        disappearOpacityAnimation.fillMode = kCAFillModeForwards
+        disappearOpacityAnimation.isRemovedOnCompletion = false
+        shapeLayer.add(disappearOpacityAnimation, forKey: "opacity")
+        
+        CATransaction.commit()
+    }
+    
+    
     // MARK: - CameraManager()
     
     fileprivate func _updateTorch(_ flashMode: CameraFlashMode) {
         captureSession?.beginConfiguration()
-        let devices = AVCaptureDevice.devices(withMediaType: AVMediaTypeVideo)
-        for  device in devices!  {
-            let captureDevice = device as! AVCaptureDevice
-            if (captureDevice.position == AVCaptureDevicePosition.back) {
-                let avTorchMode = AVCaptureTorchMode(rawValue: flashMode.rawValue)
-                if (captureDevice.isTorchModeSupported(avTorchMode!)) {
-                    do {
-                        try captureDevice.lockForConfiguration()
-                    } catch {
-                        return;
-                    }
-                    captureDevice.torchMode = avTorchMode!
-                    captureDevice.unlockForConfiguration()
+        defer { captureSession?.commitConfiguration() }
+        for captureDevice in AVCaptureDevice.videoDevices  {
+            guard captureDevice.position == AVCaptureDevicePosition.back else { continue }
+            guard let avTorchMode = AVCaptureTorchMode(rawValue: flashMode.rawValue) else { continue }
+            if (captureDevice.isTorchModeSupported(avTorchMode)) {
+                do {
+                    try captureDevice.lockForConfiguration()
+                } catch {
+                    continue
                 }
+                captureDevice.torchMode = avTorchMode
+                captureDevice.unlockForConfiguration()
             }
         }
-        captureSession?.commitConfiguration()
     }
     
     
@@ -573,47 +753,38 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
     }
     
     fileprivate func _getMovieOutput() -> AVCaptureMovieFileOutput {
-        var shouldReinitializeMovieOutput = movieOutput == nil
-        if !shouldReinitializeMovieOutput {
-            if let connection = movieOutput!.connection(withMediaType: AVMediaTypeVideo) {
-                shouldReinitializeMovieOutput = shouldReinitializeMovieOutput || !connection.isActive
+        if let movieOutput = movieOutput, let connection = movieOutput.connection(withMediaType: AVMediaTypeVideo),
+            connection.isActive {
+            return movieOutput
+        }
+        let newMoviewOutput = AVCaptureMovieFileOutput()
+        newMoviewOutput.movieFragmentInterval = kCMTimeInvalid
+        movieOutput = newMoviewOutput
+        if let captureSession = captureSession {
+            if captureSession.canAddOutput(newMoviewOutput) {
+                captureSession.beginConfiguration()
+                captureSession.addOutput(newMoviewOutput)
+                captureSession.commitConfiguration()
             }
         }
-        
-        if shouldReinitializeMovieOutput {
-            movieOutput = AVCaptureMovieFileOutput()
-            movieOutput!.movieFragmentInterval = kCMTimeInvalid
-            
-            if let captureSession = captureSession {
-                if captureSession.canAddOutput(movieOutput) {
-                    captureSession.beginConfiguration()
-                    captureSession.addOutput(movieOutput)
-                    captureSession.commitConfiguration()
-                }
-            }
-        }
-        return movieOutput!
+        return newMoviewOutput
     }
     
     fileprivate func _getStillImageOutput() -> AVCaptureStillImageOutput {
-        var shouldReinitializeStillImageOutput = stillImageOutput == nil
-        if !shouldReinitializeStillImageOutput {
-            if let connection = stillImageOutput!.connection(withMediaType: AVMediaTypeVideo) {
-                shouldReinitializeStillImageOutput = shouldReinitializeStillImageOutput || !connection.isActive
+        if let stillImageOutput = stillImageOutput, let connection = stillImageOutput.connection(withMediaType: AVMediaTypeVideo),
+            connection.isActive {
+            return stillImageOutput
+        }
+        let newStillImageOutput = AVCaptureStillImageOutput()
+        stillImageOutput = newStillImageOutput
+        if let captureSession = captureSession {
+            if captureSession.canAddOutput(newStillImageOutput) {
+                captureSession.beginConfiguration()
+                captureSession.addOutput(newStillImageOutput)
+                captureSession.commitConfiguration()
             }
         }
-        if shouldReinitializeStillImageOutput {
-            stillImageOutput = AVCaptureStillImageOutput()
-            
-            if let captureSession = captureSession {
-                if captureSession.canAddOutput(stillImageOutput) {
-                    captureSession.beginConfiguration()
-                    captureSession.addOutput(stillImageOutput)
-                    captureSession.commitConfiguration()
-                }
-            }
-        }
-        return stillImageOutput!
+        return newStillImageOutput
     }
     
     @objc fileprivate func _orientationChanged() {
@@ -625,9 +796,11 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
             currentConnection = _getMovieOutput().connection(withMediaType: AVMediaTypeVideo)
         }
         if let validPreviewLayer = previewLayer {
-            if let validPreviewLayerConnection = validPreviewLayer.connection {
-                if validPreviewLayerConnection.isVideoOrientationSupported {
-                    validPreviewLayerConnection.videoOrientation = _currentVideoOrientation()
+            if !shouldKeepViewAtOrientationChanges {
+                if let validPreviewLayerConnection = validPreviewLayer.connection {
+                    if validPreviewLayerConnection.isVideoOrientationSupported {
+                        validPreviewLayerConnection.videoOrientation = _currentVideoOrientation()
+                    }
                 }
             }
             if let validOutputLayerConnection = currentConnection {
@@ -635,11 +808,13 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
                     validOutputLayerConnection.videoOrientation = _currentVideoOrientation()
                 }
             }
-            DispatchQueue.main.async(execute: { () -> Void in
-                if let validEmbeddingView = self.embeddingView {
-                    validPreviewLayer.frame = validEmbeddingView.bounds
-                }
-            })
+            if !shouldKeepViewAtOrientationChanges {
+                DispatchQueue.main.async(execute: { () -> Void in
+                    if let validEmbeddingView = self.embeddingView {
+                        validPreviewLayer.frame = validEmbeddingView.bounds
+                    }
+                })
+            }
         }
     }
     
@@ -700,13 +875,12 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
     fileprivate func _addPreviewLayerToView(_ view: UIView) {
         embeddingView = view
         attachZoom(view)
+        attachFocus(view)
         DispatchQueue.main.async(execute: { () -> Void in
-            guard let _ = self.previewLayer else {
-                return
-            }
-            self.previewLayer!.frame = view.layer.bounds
+            guard let previewLayer = self.previewLayer else { return }
+            previewLayer.frame = view.layer.bounds
             view.clipsToBounds = true
-            view.layer.addSublayer(self.previewLayer!)
+            view.layer.addSublayer(previewLayer)
         })
     }
     
@@ -714,11 +888,11 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
         var maxZoom = CGFloat(1.0)
         beginZoomScale = CGFloat(1.0)
         
-        if cameraDevice == .back {
-            maxZoom = (backCameraDevice?.activeFormat.videoMaxZoomFactor)!
+        if cameraDevice == .back, let backCameraDevice = backCameraDevice  {
+            maxZoom = backCameraDevice.activeFormat.videoMaxZoomFactor
         }
-        else if cameraDevice == .front {
-            maxZoom = (frontCameraDevice?.activeFormat.videoMaxZoomFactor)!
+        else if cameraDevice == .front, let frontCameraDevice = frontCameraDevice {
+            maxZoom = frontCameraDevice.activeFormat.videoMaxZoomFactor
         }
         
         maxZoomScale = maxZoom
@@ -777,7 +951,12 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
                 }
             }
         case .videoOnly, .videoWithMic:
-            captureSession?.addOutput(_getMovieOutput())
+            let videoMovieOutput = _getMovieOutput()
+            if let captureSession = captureSession {
+                if captureSession.canAddOutput(videoMovieOutput) {
+                    captureSession.addOutput(videoMovieOutput)
+                }
+            }
             
             if newCameraOutputMode == .videoWithMic {
                 if let validMic = _deviceInputFromDevice(mic) {
@@ -796,7 +975,7 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
         }
         if (movieOutput == nil) {
             movieOutput = AVCaptureMovieFileOutput()
-            movieOutput!.movieFragmentInterval = kCMTimeInvalid
+            movieOutput?.movieFragmentInterval = kCMTimeInvalid
         }
         if library == nil {
             library = PHPhotoLibrary.shared()
@@ -810,10 +989,174 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
         }
     }
     
+    /**
+     Switches between the current and specified camera using a flip animation similar to the one used in the iOS stock camera app
+     */
+    
+    fileprivate var cameraTransitionView: UIView?
+    fileprivate var transitionAnimating = false
+    
+    open func _doFlipAnimation() {
+        
+        if transitionAnimating {
+            return
+        }
+        
+        if let validEmbeddingView = embeddingView {
+            if let validPreviewLayer = previewLayer {
+                
+                var tempView = UIView()
+                
+                if CameraManager._blurSupported() {
+                    
+                    let blurEffect = UIBlurEffect(style: .light)
+                    tempView = UIVisualEffectView(effect: blurEffect)
+                    tempView.frame = validEmbeddingView.bounds
+                }
+                else {
+                    
+                    tempView = UIView(frame: validEmbeddingView.bounds)
+                    tempView.backgroundColor = UIColor(white: 0.0, alpha: 0.5)
+                }
+                
+                validEmbeddingView.insertSubview(tempView, at: Int(validPreviewLayer.zPosition + 1))
+                
+                cameraTransitionView = validEmbeddingView.snapshotView(afterScreenUpdates: true)
+                
+                if let cameraTransitionView = cameraTransitionView {
+                    validEmbeddingView.insertSubview(cameraTransitionView, at: Int(validEmbeddingView.layer.zPosition + 1))
+                }
+                tempView.removeFromSuperview()
+                
+                transitionAnimating = true
+                
+                validPreviewLayer.opacity = 0.0
+                
+                DispatchQueue.main.async() {
+                    self._flipCameraTransitionView()
+                }
+            }
+        }
+    }
+    
+    // MARK: - CameraLocationManager()
+    
+    fileprivate class CameraLocationManager: NSObject, CLLocationManagerDelegate {
+        var locationManager = CLLocationManager()
+        var latestLocation: CLLocation?
+        
+        override init() {
+            super.init()
+            locationManager.delegate = self
+            locationManager.desiredAccuracy = kCLLocationAccuracyBest
+            locationManager.requestWhenInUseAuthorization()
+        }
+        
+        func startUpdatingLocation() {
+            locationManager.startUpdatingLocation()
+        }
+        
+        func stopUpdatingLocation() {
+            locationManager.stopUpdatingLocation()
+        }
+        
+        // MARK: - CLLocationManagerDelegate
+        func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+            // Pick the location with best (= smallest value) horizontal accuracy
+            latestLocation = locations.sorted { $0.horizontalAccuracy < $1.horizontalAccuracy }.first
+        }
+        
+        func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
+            if status == .authorizedAlways || status == .authorizedWhenInUse {
+                locationManager.startUpdatingLocation()
+            } else {
+                locationManager.stopUpdatingLocation()
+            }
+        }
+    }
+    
+    // Determining whether the current device actually supports blurring
+    // As seen on: http://stackoverflow.com/a/29997626/2269387
+    fileprivate class func _blurSupported() -> Bool {
+        var supported = Set<String>()
+        supported.insert("iPad")
+        supported.insert("iPad1,1")
+        supported.insert("iPhone1,1")
+        supported.insert("iPhone1,2")
+        supported.insert("iPhone2,1")
+        supported.insert("iPhone3,1")
+        supported.insert("iPhone3,2")
+        supported.insert("iPhone3,3")
+        supported.insert("iPod1,1")
+        supported.insert("iPod2,1")
+        supported.insert("iPod2,2")
+        supported.insert("iPod3,1")
+        supported.insert("iPod4,1")
+        supported.insert("iPad2,1")
+        supported.insert("iPad2,2")
+        supported.insert("iPad2,3")
+        supported.insert("iPad2,4")
+        supported.insert("iPad3,1")
+        supported.insert("iPad3,2")
+        supported.insert("iPad3,3")
+        
+        return !supported.contains(_hardwareString())
+    }
+    
+    fileprivate class func _hardwareString() -> String {
+        var name: [Int32] = [CTL_HW, HW_MACHINE]
+        var size: Int = 2
+        sysctl(&name, 2, nil, &size, &name, 0)
+        var hw_machine = [CChar](repeating: 0, count: Int(size))
+        sysctl(&name, 2, &hw_machine, &size, &name, 0)
+        
+        let hardware: String = String(cString: hw_machine)
+        return hardware
+    }
+    
+    fileprivate func _flipCameraTransitionView() {
+        
+        if let cameraTransitionView = cameraTransitionView {
+            
+            UIView.transition(with: cameraTransitionView,
+                              duration: 0.5,
+                              options: UIViewAnimationOptions.transitionFlipFromLeft,
+                              animations: nil,
+                              completion: { (finished) -> Void in
+                                self._removeCameraTransistionView()
+            })
+        }
+    }
+    
+    
+    fileprivate func _removeCameraTransistionView() {
+        
+        if let cameraTransitionView = cameraTransitionView {
+            if let validPreviewLayer = previewLayer {
+                
+                validPreviewLayer.opacity = 1.0
+            }
+            
+            UIView.animate(withDuration: 0.5,
+                           animations: { () -> Void in
+                            
+                            cameraTransitionView.alpha = 0.0
+                            
+            }, completion: { (finished) -> Void in
+                
+                self.transitionAnimating = false
+                
+                cameraTransitionView.removeFromSuperview()
+                self.cameraTransitionView = nil
+            })
+        }
+    }
+    
     fileprivate func _updateCameraDevice(_ deviceType: CameraDevice) {
         if let validCaptureSession = captureSession {
             validCaptureSession.beginConfiguration()
-            let inputs = validCaptureSession.inputs as! [AVCaptureInput]
+            defer { validCaptureSession.commitConfiguration() }
+            let inputs: [AVCaptureInput] = validCaptureSession.inputs?.flatMap { $0 as? AVCaptureInput } ?? []
             
             for input in inputs {
                 if let deviceInput = input as? AVCaptureDeviceInput {
@@ -842,29 +1185,57 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
                     }
                 }
             }
-            validCaptureSession.commitConfiguration()
         }
     }
     
     fileprivate func _updateFlasMode(_ flashMode: CameraFlashMode) {
         captureSession?.beginConfiguration()
-        let devices = AVCaptureDevice.devices(withMediaType: AVMediaTypeVideo)
-        for  device in devices!  {
-            let captureDevice = device as! AVCaptureDevice
+        defer { captureSession?.commitConfiguration() }
+        for captureDevice in AVCaptureDevice.videoDevices  {
             if (captureDevice.position == AVCaptureDevicePosition.back) {
-                let avFlashMode = AVCaptureFlashMode(rawValue: flashMode.rawValue)
-                if (captureDevice.isFlashModeSupported(avFlashMode!)) {
+                guard let avFlashMode = AVCaptureFlashMode(rawValue: flashMode.rawValue) else { continue }
+                if (captureDevice.isFlashModeSupported(avFlashMode)) {
                     do {
                         try captureDevice.lockForConfiguration()
                     } catch {
                         return
                     }
-                    captureDevice.flashMode = avFlashMode!
+                    captureDevice.flashMode = avFlashMode
                     captureDevice.unlockForConfiguration()
                 }
             }
         }
-        captureSession?.commitConfiguration()
+    }
+    
+    fileprivate func _performShutterAnimation(_ completion: (() -> Void)?) {
+        
+        if let validPreviewLayer = previewLayer {
+            
+            DispatchQueue.main.async {
+                
+                let duration = 0.1
+                
+                CATransaction.begin()
+                
+                if let completion = completion {
+                    
+                    CATransaction.setCompletionBlock(completion)
+                }
+                
+                let fadeOutAnimation = CABasicAnimation(keyPath: "opacity")
+                fadeOutAnimation.fromValue = 1.0
+                fadeOutAnimation.toValue = 0.0
+                validPreviewLayer.add(fadeOutAnimation, forKey: "opacity")
+                
+                let fadeInAnimation = CABasicAnimation(keyPath: "opacity")
+                fadeInAnimation.fromValue = 0.0
+                fadeInAnimation.toValue = 1.0
+                fadeInAnimation.beginTime = CACurrentMediaTime() + duration * 2.0
+                validPreviewLayer.add(fadeInAnimation, forKey: "opacity")
+                
+                CATransaction.commit()
+            }
+        }
     }
     
     fileprivate func _updateCameraQualityMode(_ newCameraOutputQuality: CameraOutputQuality) {
@@ -927,6 +1298,14 @@ open class CameraManager: NSObject, AVCaptureFileOutputRecordingDelegate, UIGest
     
     deinit {
         stopAndRemoveCaptureSession()
-        _stopFollowingDeviceOrientation()        
+        _stopFollowingDeviceOrientation()
+    }
+}
+
+
+fileprivate extension AVCaptureDevice {
+    fileprivate static var videoDevices: [AVCaptureDevice] {
+        guard let devices = AVCaptureDevice.devices(withMediaType: AVMediaTypeVideo) else { return [] }
+        return devices.flatMap { $0 as? AVCaptureDevice }
     }
 }